Part1
Part 1 is a simple version of the game. When the file is loaded a color is randomly selected from an array of colors and a loop begins which repeatedly asks the player what the color is. However, no feedback is given to the player depending on what he/she enters. The loop simply finishes when the player enters the correct color. Make sure you save a copy of this as, for example, part1.html, before you move on to the next part.